damn...

+++ navy-apps/tests/dummy/dummy.c
@@ -5,7 +5,7 @@
 #endif
 
 #define SYS_yield 1
-extern intptr_t _syscall_(intptr_t, intptr_t, intptr_t, intptr_t);
+extern int _syscall_(int, uintptr_t, uintptr_t, uintptr_t);
 
 int main() {
   return _syscall_(SYS_yield, 0, 0, 0);
